requirements.txt
# Core LangChain ecosystem
langchain>=0.2.14
langgraph>=0.1.3
langchain-openai>=0.1.7

# Utilities
python-dotenv>=1.0.1
typing-extensions>=4.12.2

# Optional (for better console experience and debugging)
rich>=13.7.1
ipython>=8.26.0
